Move scipyfit tests to new file
